+++ Lab5/RectangleLawn.cpp
@@ -18,10 +18,10 @@ namespace lab5
 
 	unsigned int RectangleLawn::GetFencePrice(eFenceType fenceType) const
 	{
-		unsigned int meter = (unsigned int) (GetMinimumFencesCount() * 0.25);
+		double meter = GetMinimumFencesCount() * 0.25;
 		unsigned int price = fenceType == lab5::RED_CEDAR ? 6 : 7;
 
-		return meter * price;
+		return (unsigned int)meter * price;
 	}
 
 	RectangleLawn::RectangleLawn(unsigned int width, unsigned height)
